+++ src/ts/app/interfaces/modal.ts
@@ -83,6 +83,7 @@ export interface ModalParams {
     delayInMilliSeconds?: number,
     backDrop?: BackDropParams;
     disableEscKey?: boolean;
+    closeOnOutsideMouseClick? : boolean;
 }
 
 export interface BackDropParams {